TableLayout: cleaned up code
